Refactor team lifecycle tool facade
